move test_join_group_then_handshake
