+++ nodes/key-vault/src/tests.rs
@@ -1,7 +1,7 @@
 use crate::{handlers::handle_health_check, Server as KeyVaultServer};
 use actix_web::{http::StatusCode, test, web, App};
 use anonify_eth_driver::utils::*;
-use frame_common::crypto::Ed25519ChallengeResponse;
+use frame_common::crypto::{AccountId, Ed25519ChallengeResponse};
 use frame_config::{ANONIFY_ABI_PATH, FACTORY_ABI_PATH, PJ_ROOT_DIR};
 use frame_host::EnclaveDir;
 use frame_sodium::{SodiumCiphertext, SodiumPubKey};
@@ -9,8 +9,10 @@ use once_cell::sync::Lazy;
 use rand_core::{CryptoRng, RngCore};
 use serde_json::json;
 use state_runtime_node_server::{handlers::*, Server as ERC20Server};
-use std::{env, fs, path::Path, str::FromStr, sync::Arc};
+use std::{env, fs, path::Path, str::FromStr, sync::Arc, time};
 use web3::{contract::Options, types::Address};
+
+const SYNC_TIME: u64 = 1500;
 
 #[actix_rt::test]
 async fn test_health_check() {
@@ -44,6 +46,175 @@ async fn test_health_check() {
     let req = test::TestRequest::get().uri("/api/v1/health").to_request();
     let resp = test::call_service(&mut healthy_app, req).await;
     assert_eq!(resp.status(), StatusCode::OK);
+}
+
+#[actix_rt::test]
+async fn test_join_group_then_handshake() {
+    set_env_vars();
+
+    let eth_url = env::var("ETH_URL").expect("ETH_URL is not set");
+
+    // Setup key-vault server
+    let key_vault_server_enclave = EnclaveDir::new()
+        .init_enclave(true)
+        .expect("Failed to initialize server enclave.");
+    let key_vault_server_eid = key_vault_server_enclave.geteid();
+    let key_vault_server = Arc::new(KeyVaultServer::new(key_vault_server_eid).run().await);
+    let _key_vault_app = test::init_service(App::new().data(key_vault_server.clone())).await;
+    std::thread::sleep(std::time::Duration::from_secs(1));
+
+    // Enclave must be initialized in main function.
+    let enclave1 = EnclaveDir::new()
+        .init_enclave(true)
+        .expect("Failed to initialize enclave.");
+    let eid1 = enclave1.geteid();
+    // just for testing
+    let mut csprng = rand::thread_rng();
+
+    let server1 = ERC20Server::new(eid1).await.run().await;
+    let server1 = Arc::new(server1);
+    let mut app1 = test::init_service(
+        App::new()
+            .data(server1.clone())
+            .route("/api/v1/state", web::get().to(handle_get_state))
+            .route(
+                "/api/v1/enclave_encryption_key",
+                web::get().to(handle_enclave_encryption_key),
+            ),
+    )
+    .await;
+
+    let enclave2 = EnclaveDir::new()
+        .init_enclave(true)
+        .expect("Failed to initialize enclave.");
+    let eid2 = enclave2.geteid();
+    let server2 = ERC20Server::new(eid2).await.run().await;
+    let server2 = Arc::new(server2);
+    let mut app2 = test::init_service(
+        App::new()
+            .data(server2.clone())
+            .route("/api/v1/state", web::post().to(handle_send_command))
+            .route("/api/v1/state", web::get().to(handle_get_state))
+            .route("/api/v1/key_rotation", web::post().to(handle_key_rotation))
+            .route(
+                "/api/v1/enclave_encryption_key",
+                web::get().to(handle_enclave_encryption_key),
+            ),
+    )
+    .await;
+
+    // Party 1
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/enclave_encryption_key")
+        .to_request();
+    let resp = test::call_service(&mut app1, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let enc_key_resp: state_runtime_node_api::enclave_encryption_key::get::Response =
+        test::read_body_json(resp).await;
+    let enc_key1 = verify_enclave_encryption_key(
+        enc_key_resp.enclave_encryption_key,
+        &*FACTORY_ABI_PATH,
+        &*ANONIFY_ABI_PATH,
+        &eth_url,
+    )
+    .await;
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key1))
+        .to_request();
+    let resp = test::call_service(&mut app1, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let balance: state_runtime_node_api::state::get::Response = test::read_body_json(resp).await;
+    assert_eq!(balance.state, 0);
+
+    // Party 2
+
+    other_turn();
+
+    // using the same encryption key because app2 have to sync with bc, but app2's enclave encryption key cannot be set here
+    // so using app1's key
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key1))
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp); // return 200 OK: becuse allowed same enclave encryption key
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/enclave_encryption_key")
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let enc_key_resp: state_runtime_node_api::enclave_encryption_key::get::Response =
+        test::read_body_json(resp).await;
+    let enc_key = verify_enclave_encryption_key(
+        enc_key_resp.enclave_encryption_key,
+        &*FACTORY_ABI_PATH,
+        &*ANONIFY_ABI_PATH,
+        &eth_url,
+    )
+    .await;
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key))
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let balance: state_runtime_node_api::state::get::Response = test::read_body_json(resp).await;
+    assert_eq!(balance.state, 0);
+
+    let init_100_req = init_100_req(&mut csprng, &enc_key, 1);
+    let req = test::TestRequest::post()
+        .uri("/api/v1/state")
+        .set_json(&init_100_req)
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key))
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let balance: state_runtime_node_api::state::get::Response = test::read_body_json(resp).await;
+    assert_eq!(balance.state, 100);
+
+    let req = test::TestRequest::post()
+        .uri("/api/v1/key_rotation")
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    actix_rt::time::delay_for(time::Duration::from_millis(SYNC_TIME)).await;
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key))
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let balance: state_runtime_node_api::state::get::Response = test::read_body_json(resp).await;
+    assert_eq!(balance.state, 100);
+
+    let transfer_10_req = transfer_10_req(&mut csprng, &enc_key, 2);
+    let req = test::TestRequest::post()
+        .uri("/api/v1/state")
+        .set_json(&transfer_10_req)
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+
+    let req = test::TestRequest::get()
+        .uri("/api/v1/state")
+        .set_json(&balance_of_req(&mut csprng, &enc_key))
+        .to_request();
+    let resp = test::call_service(&mut app2, req).await;
+    assert!(resp.status().is_success(), "response: {:?}", resp);
+    let balance: state_runtime_node_api::state::get::Response = test::read_body_json(resp).await;
+    assert_eq!(balance.state, 90);
 }
 
 #[actix_rt::test]
@@ -689,6 +860,13 @@ async fn verify_enclave_encryption_key<P: AsRef<Path> + Copy>(
     enclave_encryption_key
 }
 
+fn other_turn() {
+    env::remove_var("MY_ROSTER_IDX");
+    env::remove_var("ACCOUNT_INDEX");
+    env::set_var("MY_ROSTER_IDX", "1");
+    env::set_var("ACCOUNT_INDEX", "1");
+}
+
 fn init_100_req<CR>(
     csprng: &mut CR,
     enc_key: &SodiumPubKey,
@@ -722,6 +900,51 @@ where
         "counter": counter,
     });
 
+    let ciphertext =
+        SodiumCiphertext::encrypt(csprng, &enc_key, &serde_json::to_vec(&req).unwrap()).unwrap();
+
+    state_runtime_node_api::state::post::Request {
+        ciphertext,
+        user_id: None,
+    }
+}
+
+fn transfer_10_req<CR>(
+    csprng: &mut CR,
+    enc_key: &SodiumPubKey,
+    counter: u32,
+) -> state_runtime_node_api::state::post::Request
+where
+    CR: RngCore + CryptoRng,
+{
+    let sig = [
+        227, 77, 52, 167, 149, 64, 24, 23, 103, 227, 13, 120, 90, 186, 1, 62, 110, 60, 186, 247,
+        143, 247, 19, 71, 85, 191, 224, 5, 38, 219, 96, 44, 196, 154, 181, 50, 99, 58, 20, 125,
+        244, 172, 212, 166, 234, 203, 208, 77, 9, 232, 77, 248, 152, 81, 106, 49, 120, 34, 212, 89,
+        92, 100, 221, 14,
+    ];
+    let pubkey = [
+        164, 189, 195, 42, 48, 163, 27, 74, 84, 147, 25, 254, 16, 14, 206, 134, 153, 148, 33, 189,
+        55, 149, 7, 15, 11, 101, 106, 28, 48, 130, 133, 143,
+    ];
+    let challenge = [
+        157, 61, 16, 189, 40, 124, 88, 101, 19, 36, 155, 229, 245, 123, 189, 124, 222, 114, 215,
+        186, 25, 30, 135, 114, 237, 169, 138, 122, 81, 61, 43, 183,
+    ];
+    let access_policy = Ed25519ChallengeResponse::new_from_bytes(sig, pubkey, challenge);
+
+    let req = json!({
+        "access_policy": access_policy,
+        "runtime_params": {
+            "amount": 10,
+            "recipient": AccountId([
+                236, 126, 92, 200, 50, 125, 9, 112, 74, 58, 35, 60, 181, 105, 198, 107, 62, 111, 168,
+                118,
+            ])
+        },
+        "cmd_name": "transfer",
+        "counter": counter,
+    });
     let ciphertext =
         SodiumCiphertext::encrypt(csprng, &enc_key, &serde_json::to_vec(&req).unwrap()).unwrap();
